gastronom/requirements.txt
Django==2.2
django-jet==1.0.8
djangorestframework==3.12.2
django-super-inlines==0.1.4
python-telegram-bot==13.0
python-dateutil==2.8.1
Pillow==7.2.0
django-tinymce==3.2.0
celery==5.0.3
django-imagekit
django-filter==2.4.0
django-celery-results==2.0.0
sentry-sdk==0.19.5
